requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
comtypes==1.4.11
filelock==3.18.0
fsspec==2025.5.1
huggingface-hub==0.32.1
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
packaging==25.0
plyer==2.1.0
PyAudio==0.2.14
pygame==2.6.1
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyttsx3==2.98
pywin32==310
PyYAML==6.0.2
redis==6.1.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==80.8.0
six==1.17.0
spotipy==2.25.1
sympy==1.14.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
typing_extensions==4.13.2
urllib3==2.4.0
wheel==0.45.1
whisper==1.1.10
zstandard @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_65utj9q8ya/croot/zstandard_1731360545821/work
spotipy
spotipy


# 🌐 FastAPI Backend
# FastAPI Backend
fastapi
uvicorn
python-multipart

# 🔊 Voice Input and Audio Playback
pygame==2.6.1
sounddevice==0.4.6

# 🧠 Speech Recognition
ffmpeg-python==0.2.0
git+https://github.com/openai/whisper.git

# 🤖 NLP & Intent Detection
# Voice Input and Audio Playback
pygame==2.6.1
sounddevice==0.4.6
pyaudio

# Speech Recognition and Text-to-Speech
ffmpeg-python==0.2.0


# NLP & Intent Detection
transformers==4.40.0
torch==2.3.0
wit==4.2.0
openai

# 🎵 Spotify API
spotipy==2.23.0

# 📦 Environment Variables
python-dotenv==1.0.1

# 🧮 Data Handling & Utilities
numpy==1.26.4
requests==2.31.0
pymongo==4.6.3

# 📅 Calendar & External API Integration
google-api-python-client==2.122.0

# 🧪 Testing and Debugging (Optional)
ipython==8.24.0
rich==13.7.1
scikit-learn

# Spotify API
spotipy==2.25.1

# Environment Variables
python-dotenv==1.0.1

# Data Handling & Utilities
numpy==1.26.4
requests==2.31.0
pymongo==4.6.31
python-dateutil

# Calendar & External API Integration
google-api-python-client==2.122.0

# Notifications
plyer

# Testing and Debugging (Optional)
ipython==8.24.0
rich==13.7.1spotipy
flask
elevenlabs